refactor: 스토리북, 컴포넌트, 페이지에 대한 폴더 구조 만들어 관리
