+++ docs/REFACTOR.md
@@ -33,5 +33,5 @@
 - [ ] window.location.hash를 통해 url 초기 설정
 - [ ] 모바일에서 화면이 잘 보이도록 반응형으로 만드는 기능 구현 - 현재 볼더가 페이지에 적용되어 있는 것을 해제한다.
 - [ ] 홈페이지에서 보유 카드 타이틀 옆에 현재 카드 개수를 보여주며 전부 삭제 기능을 구현한다.
-- [ ] 스토리북 ,컴포넌트, 페이지에 대한 폴더 만들어 정리하기
+- [x] 스토리북 ,컴포넌트, 페이지에 대한 폴더 만들어 정리하기
 - [ ] 스토리 북에서 컴포넌트의 종류를 컴포넌트 단위로 나누기 -> 더 다양한 스토리를 주기
