+++ src/App.tsx
@@ -1,10 +1,11 @@
-import { FormEvent, useState } from "react";
-import AddCardPage from "./components/AddCardPage";
-import Homepage from "./components/Homepage";
-import { ADD_CARD_PAGE, HOME_PAGE, LOCAL_STORAGE_CARD_KEY } from "./constant";
-import { useLocalStorage } from "./hooks/useLocalStorage";
 import "./styles/index.css";
+
+import { FormEvent, useState } from "react";
+import { ADD_CARD_PAGE, HOME_PAGE, LOCAL_STORAGE_CARD_KEY } from "./constant";
 import { CardInfo, PageInfo } from "./types";
+import { useLocalStorage } from "./hooks/useLocalStorage";
+import Homepage from "./pages/HomePage";
+import AddCardPage from "./pages/AddCardPage";
 
 export default function App() {
   const [pageIndex, setPageIndex] = useState<PageInfo>(HOME_PAGE);
