+++ src/components/card/input/CardNumberInput.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
-import { UseInputProps } from "../hooks/useInput";
-import Input from "./Input";
+import Input from "../../common/Input";
+import { UseInputProps } from "../../../hooks/useInput";
 
 interface CardNumberInputProps {
   firstNumber: UseInputProps;
