+++ src/components/card/input/ExpiracyInput.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
-import { UseInputProps } from "../hooks/useInput";
-import Input from "./Input";
-import ToolTip from "./ToolTip";
+import { UseInputProps } from "../../../hooks/useInput";
+import Input from "../../common/Input";
+import ToolTip from "../../common/ToolTip";
 
 interface ExpiracyInputProps {
   year: UseInputProps;
