+++ src/hooks/useLocalStorage.tsx
@@ -14,7 +14,7 @@ export const useLocalStorage = <T,>(
   useEffect(() => {
     const savedData = JSON.parse(localStorage.getItem(key) || "[]");
     if (savedData && savedData.length > 0) setData(savedData);
-  }, []);
+  }, [key]);
 
   return [value, setValue];
 };
