+++ src/stories/common/InformationButton.stories.tsx
@@ -1,6 +1,5 @@
 import { Meta, StoryObj } from "@storybook/react";
-
-import InformationButton from "../components/InformationButton";
+import InformationButton from "../../components/common/InformationButton";
 
 const meta: Meta<typeof InformationButton> = {
   component: InformationButton,
