+++ src/stories/pages/AddCardPage.stories.tsx
@@ -1,6 +1,5 @@
 import { Meta, StoryObj } from "@storybook/react";
-
-import AddCardPage from "../components/AddCardPage";
+import AddCardPage from "../../pages/AddCardPage";
 
 const meta: Meta<typeof AddCardPage> = {
   component: AddCardPage,
@@ -12,5 +11,8 @@ type Story = StoryObj<typeof AddCardPage>;
 
 //title,cardNumberSet,owner,expiracy
 export const AddCard: Story = {
-  args: {},
+  args: {
+    onClick: () => {},
+    onSubmit: () => {},
+  },
 };
