+++ src/stories/pages/Homepage.stories.tsx
@@ -1,6 +1,5 @@
 import { Meta, StoryObj } from "@storybook/react";
-
-import Homepage from "../components/Homepage";
+import Homepage from "../../pages/HomePage";
 
 const meta: Meta<typeof Homepage> = {
   component: Homepage,
